+++ test/bigtest/network/factories/item.js
@@ -1,4 +1,4 @@
-import { faker } from '@bigtest/mirage';
+import faker from 'faker';
 
 import Factory from './application';